supabase/Api-postgre/Supabase/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
pydantic==2.5.2
psycopg2-binary==2.9.10
